+++ tests/cmake/psa_tests.cmake
@@ -1,0 +1,14 @@
+set(PSA_CONFIG_FILE psa_config.txt)
+set(PSA_TEST_CMD ${CMAKE_INSTALL_PREFIX}/${SMW_TESTS_TARGET_SCRIPTS_DIR}/run_psa_test.sh)
+set(PSA_COMMAND ${PSA_TEST_CMD} ${PSA_CONFIG_FILE})
+
+# Install config file
+install(FILES ${SMW_CONFIG_SRC_DIR}/${PSA_CONFIG_FILE}
+	DESTINATION ${SMW_TESTS_TARGET_CONFIG_DIR}
+	EXCLUDE_FROM_ALL
+	COMPONENT ${PROJECT_NAME})
+
+# Get all PSA test definition files
+FILE(GLOB PSA_TESTS ${TEST_DEF_SRC_DIR}/*_PSA_*.json)
+
+add_and_install_tests("${PSA_TESTS}" "${PSA_COMMAND}")